본인의 게시글만 수정/삭제 가능 처리

+++ src/main/java/com/example/prj1/board/service/BoardService.java
@@ -83,20 +83,34 @@ public class BoardService {
         return dto;
     }
 
-    public void remove(Integer id) {
-        boardRepository.deleteById(id);
+    public Boolean remove(Integer id, MemberDto user) {
+        if (user != null) {
+            Member db = boardRepository.findById(id).get().getWriter();
+
+            if (db.getId().equals(user.getId())) {
+                boardRepository.deleteById(id);
+                return true;
+            }
+        }
+        return false;
     }
 
-    public void update(BoardForm data) {
-        // 조회
-        Board board = boardRepository.findById(data.getId()).get();
-        // 수정
-        board.setTitle(data.getTitle());
-        board.setContent(data.getContent());
-//        board.setWriter(data.getWriter());
+    public boolean update(BoardForm data, MemberDto user) {
+        if (user != null) {
+            // 조회
+            Board board = boardRepository.findById(data.getId()).get();
 
-        // 저장
-        boardRepository.save(board);
+            if (board.getWriter().getId().equals(user.getId())) {
+                // 수정
+                board.setTitle(data.getTitle());
+                board.setContent(data.getContent());
+                //        board.setWriter(data.getWriter());
 
+                // 저장
+                boardRepository.save(board);
+                return true;
+            }
+        }
+        return false;
     }
 }
